pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "searcharena"
version = "0.1.0"
description = "SearchArena environment for OpenEnv"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "openenv-core[core]>=0.2.2",
    "openai>=2.31.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=4.0.0",
]
datagen = [
    "anthropic>=0.75.0",
    "chromadb>=1.4.0",
    "tiktoken>=0.12.0",
    "rich>=14.3.4",
    "requests>=2.32.5",
]
datagen-web = [
    "searcharena[datagen]",
]
datagen-sec = [
    "searcharena[datagen]",
]
datagen-patents = [
    "searcharena[datagen]",
    "python-docx>=1.2.0",
    "pandas>=2.0.0",
    "httpx[http2]>=0.28.1",
]

[project.scripts]
server = "searcharena.server.app:main"
inference = "searcharena.cli.inference:main"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["searcharena*"]

[tool.setuptools.package-data]
"searcharena.tasks.resources.sample" = ["*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
